+++ packages/native/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @regenrek/gpuix-native
 
+## 0.5.0-regenrek.2
+
 ## 0.5.0-regenrek.1
 
 ### Patch Changes